twister: split report classes out
Move report classes out into own file and split out of testplan class.
